url issues fixed

+++ wp-content/themes/twentytwelve/classes/user.class.php
@@ -97,7 +97,7 @@ class User
 
         //all the rules defined//
         $v->rule('required', ['gender', 'xooma_member_id','birth_date']);
-        $v->rule('integer', ['phone_no','xooma_member_id']);
+        $v->rule('numeric', ['phone_no','xooma_member_id']);
         $v->rule('equalTo', 'gender', 'male');
         $v->rule('fixedLength', 'xooma_member_id', 6);
         $v->rule('date', 'birth_date');
